restructured the codebase
