+++ results/emotion_detection_data.csv
@@ -1,2 +1,2 @@
 Person ID,Angry (s),Sad (s),Happy (s),Fear (s),Disgust (s),Neutral (s),Surprise (s)
-Person 1,0,8.281079053878784,9.729175567626953,0,0,0.5926809310913086,0
+Person 1,0,9.791351795196533,1.2825520038604736,0,0,0.6334075927734375,0
