+++ results/eye_tracking_data.csv
@@ -1,2 +1,2 @@
 Person ID,Duration Eyes Closed (s),Duration Looking Left (s),Duration Looking Right (s),Duration Looking Straight (s),Left Counts,Right Counts,Straight Counts
-Person 1,1.400000000000001,0,0,1.5333333333333348,0,0,42
+Person 1,0.4333333333333333,0,0.03333333333333333,1.2000000000000004,0,1,13
