+++ results/head_pose_data.csv
@@ -1,2 +1,2 @@
 Person ID,Looking Forward (s),Looking Left (s),Looking Right (s),Looking Up (s),Looking Down (s)
-Person 1,1.9000000000000026,0.2333333333333333,0.1,0.7,0
+Person 1,1.0666666666666667,0,0.13333333333333333,0.39999999999999997,0
